Improve editable outline feature
