+++ mainapp.py
@@ -6,6 +6,7 @@ import hashlib
 from app.parse_input_files import extract_text_and_tokens
 from app.generate_llm_output import generate_outline, check_alignment, generate_questions, set_runtime_config
 from app.export_docx import build_questions_docx
+from app.display_content import display_editable_outline, display_static_outline
 from app.save_load_progress import save_load_panel, apply_pending_restore
 import app.constants as const
 
@@ -49,12 +50,11 @@ def _sig_questions(questions_by_lo: Dict[str, list]) -> str:
     payload = "\n".join(parts)
     return hashlib.sha1(payload.encode("utf-8")).hexdigest()
 
-
 ################################################
 # App setup
 ################################################
 # Page config
-st.set_page_config(page_title="ALTO - Design", page_icon=":mortar_board:", layout="wide", initial_sidebar_state="collapsed")
+st.set_page_config(page_title="BEACON - Design", page_icon=":mortar_board:", layout="wide", initial_sidebar_state="collapsed")
 
 # Apply any pending restore from saved session state
 apply_pending_restore()
@@ -92,7 +92,7 @@ set_runtime_config(ss["MOCK_MODE"], ss["OPENAI_MODEL"])
 
 # Title and warning based on current mock setting
 mock_warning = "   :red[⚠️ MOCK MODE is ON]"
-st.title(f":mortar_board: ALTO - Design{mock_warning if ss['MOCK_MODE'] else ''}")
+st.title(f":mortar_board: BEACON - Design{mock_warning if ss['MOCK_MODE'] else ''}")
 st.markdown("##### _AI support for smarter course design._")
 
 # --------------------------------------------------------------
@@ -233,8 +233,7 @@ Export questions to Microsoft Word
 
     # crisp separation from the rest of the page
     st.divider()
-    # small bottom breathing room
-    #st.write("")
+
 
 ################################################
 # 1 Course Outline
@@ -324,196 +323,16 @@ Investing time upfront in the outline will make the presentation of content more
             _clear_outline_widget_state()
             ss['generated_outline'] = generate_outline(ss["outline_guidance"].strip(), ss["course_text"])
 
-    def _get_outline_node(path_parts):
-        """Return the container and final key/index for a dotted outline path."""
-        if "generated_outline" not in ss:
-            raise KeyError("Outline is not available in session state.")
-
-        if not path_parts:
-            raise ValueError("Path cannot be empty.")
-
-        node: Any = ss["generated_outline"]
-        for idx, part in enumerate(path_parts[:-1]):
-            next_part = path_parts[idx + 1]
-            if isinstance(node, list):
-                index = int(part)
-                node = node[index]
-            elif isinstance(node, dict):
-                if part not in node:
-                    node[part] = [] if next_part.isdigit() else {}
-                node = node[part]
-            else:
-                raise KeyError(f"Unsupported path segment '{part}' for node type {type(node)}")
-
-        return node, path_parts[-1]
-
-    def _normalize_outline_input(path: str, raw_value: str, existing: Any):
-        if isinstance(existing, list) or path.endswith("keyPoints"):
-            lines = [line.strip() for line in (raw_value or "").splitlines()]
-            return [line for line in lines if line]
-        return raw_value
-
-    def _update_outline_value(path: str, widget_key: str):
-        if "generated_outline" not in ss:
-            return
-
-        raw_value = ss.get(widget_key, "")
-        path_parts = path.split(".")
-        parent, final_key = _get_outline_node(path_parts)
-
-        if isinstance(parent, list):
-            index = int(final_key)
-            existing = parent[index] if index < len(parent) else None
-            normalized = _normalize_outline_input(path, raw_value, existing)
-            if index < len(parent):
-                parent[index] = normalized
-            else:
-                parent.append(normalized)
-        else:
-            existing = parent.get(final_key)
-            normalized = _normalize_outline_input(path, raw_value, existing)
-            parent[final_key] = normalized
-
-        ss[widget_key] = "\n".join(normalized) if isinstance(normalized, list) else normalized
-
-    def outline_text_field(label: str, path: str, value: Any, *, area: bool = False):
-        widget_key = f"outline__{path.replace('.', '__')}"
-
-        if isinstance(value, list):
-            initial_value = "\n".join(value)
-        else:
-            initial_value = value or ""
-
-        if widget_key not in ss:
-            ss[widget_key] = initial_value
-
-        field_kwargs = {
-            "label": label,
-            "key": widget_key,
-            "on_change": _update_outline_value,
-            "args": (path, widget_key),
-        }
-
-        if area:
-            st.text_area(**field_kwargs)
-        else:
-            st.text_input(**field_kwargs)
-
-    def display_outline(outline: Dict[str, Any]):
-        """Render the editable outline using text inputs backed by session state."""
-
-        outline.setdefault("courseLevelObjectives", [])
-        outline.setdefault("modules", [])
-
-        st.header("Course Outline")
-        outline_text_field("Course title", "courseTitle", outline.get("courseTitle", ""))
-
-        st.subheader("Course-Level Objectives")
-        course_objectives = outline["courseLevelObjectives"]
-        for idx, objective in enumerate(course_objectives):
-            outline_text_field(f"Objective {idx + 1}", f"courseLevelObjectives.{idx}", objective)
-
-        if st.button("Add course objective", key="outline_add_course_objective"):
-            course_objectives.append("")
-            new_path = f"courseLevelObjectives.{len(course_objectives) - 1}"
-            ss[f"outline__{new_path.replace('.', '__')}"] = ""
-            st.rerun()
-
-        st.divider()
-
-        for module_index, module in enumerate(outline["modules"]):
-            module.setdefault("sections", [])
-            st.markdown(f"### Module {module_index + 1}")
-            outline_text_field(
-                "Module title",
-                f"modules.{module_index}.moduleTitle",
-                module.get("moduleTitle", ""),
-            )
-            outline_text_field(
-                "Module overview",
-                f"modules.{module_index}.overview",
-                module.get("overview", ""),
-                area=True,
-            )
-
-            for section_index, section in enumerate(module["sections"]):
-                section.setdefault("sectionLevelObjectives", [])
-                section.setdefault("units", [])
-
-                st.markdown(f"#### Section {module_index + 1}.{section_index + 1}")
-                outline_text_field(
-                    "Section title",
-                    f"modules.{module_index}.sections.{section_index}.sectionTitle",
-                    section.get("sectionTitle", ""),
-                )
-
-                st.caption("Section-level objectives")
-                for obj_index, section_obj in enumerate(section["sectionLevelObjectives"]):
-                    cols = st.columns([1, 3])
-                    with cols[0]:
-                        outline_text_field(
-                            "Bloom level",
-                            f"modules.{module_index}.sections.{section_index}.sectionLevelObjectives.{obj_index}.bloomsLevel",
-                            section_obj.get("bloomsLevel", ""),
-                        )
-                    with cols[1]:
-                        outline_text_field(
-                            "Objective text",
-                            f"modules.{module_index}.sections.{section_index}.sectionLevelObjectives.{obj_index}.objectiveText",
-                            section_obj.get("objectiveText", ""),
-                            area=True,
-                        )
-
-                add_section_obj_key = f"outline_add_section_obj_{module_index}_{section_index}"
-                if st.button("Add section objective", key=add_section_obj_key):
-                    section["sectionLevelObjectives"].append({"bloomsLevel": "", "objectiveText": ""})
-                    base_path = f"modules.{module_index}.sections.{section_index}.sectionLevelObjectives.{len(section['sectionLevelObjectives']) - 1}"
-                    ss[f"outline__{base_path.replace('.', '__') + '__bloomsLevel'}"] = ""
-                    ss[f"outline__{base_path.replace('.', '__') + '__objectiveText'}"] = ""
-                    st.rerun()
-
-                for unit_index, unit in enumerate(section["units"]):
-                    unit.setdefault("unitLevelObjective", {})
-                    unit.setdefault("keyPoints", [])
-
-                    with st.expander(
-                        f"Unit {module_index + 1}.{section_index + 1}.{unit_index + 1}: {unit.get('unitTitle', 'Untitled unit') or 'Untitled unit'}",
-                        expanded=False,
-                    ):
-                        outline_text_field(
-                            "Unit title",
-                            f"modules.{module_index}.sections.{section_index}.units.{unit_index}.unitTitle",
-                            unit.get("unitTitle", ""),
-                        )
-
-                        cols = st.columns([1, 3])
-                        with cols[0]:
-                            outline_text_field(
-                                "Bloom level",
-                                f"modules.{module_index}.sections.{section_index}.units.{unit_index}.unitLevelObjective.bloomsLevel",
-                                unit["unitLevelObjective"].get("bloomsLevel", ""),
-                            )
-                        with cols[1]:
-                            outline_text_field(
-                                "Objective text",
-                                f"modules.{module_index}.sections.{section_index}.units.{unit_index}.unitLevelObjective.objectiveText",
-                                unit["unitLevelObjective"].get("objectiveText", ""),
-                                area=True,
-                            )
-
-                        outline_text_field(
-                            "Key points",
-                            f"modules.{module_index}.sections.{section_index}.units.{unit_index}.keyPoints",
-                            "\n".join(unit.get("keyPoints", [])),
-                            area=True,
-                        )
-
     # --- Display Output ---
     if 'generated_outline' in ss:
-        st.success("Course outline generated successfully!")
         
+        # Switch between static and editable outline view
+        st.toggle("Editable outline", key="editable_outline", value=False)
         # Display the formatted outline
-        display_outline(ss['generated_outline'])
+        if ss["editable_outline"]:
+            display_editable_outline(ss['generated_outline'])
+        else:
+            display_static_outline(ss['generated_outline'])
 
     # --- Navigation ---
     st.divider()
@@ -558,10 +377,8 @@ def render_step_2():
         if ss.get("module_sig") and new_mod_sig != ss["module_sig"]:
             st.toast("Module content changed — LOs and questions cleared.")
             clear_module_dependent_outputs()
-            #st.caption(f"old sig: {ss['module_sig'][:8]}… → new: {new_mod_sig[:8]}…")
 
         # Persist the latest parse & signature
-        #ss["processed_file_keys"] = current_file_keys     # keep if you like, but no longer used to clear
         ss["module_files"] = [f.name for f in files]
         ss["module_text"] = text
         ss["module_tokens"] = tokens
@@ -574,14 +391,12 @@ def render_step_2():
     # Display currently uploaded files from the session state (stable across reruns)
     if ss["module_files"]:
         st.caption("Currently uploaded files (To change, use file picker above):")
-        #current_files = "\n".join([f"{i+1}. {f.name}" for i, f in enumerate(ss["module_files"])])
         current_files = "\n".join([f"{i+1}. {fname}" for i, fname in enumerate(ss["module_files"])])
         st.markdown(current_files)
 
     # Display token count & preview from session (stable across reruns)
     st.caption(f"Estimated tokens: {ss.get('module_tokens', 0):,}")
     with st.expander("Preview first 5,000 characters", expanded=False):
-        #st.text_area("Preview", (ss.get("module_text") or "")[:5000], height=150, disabled=True, key="preview_area")
         st.text_area("Preview", (ss.get("module_text") or "")[:5000], height=150, disabled=True, label_visibility="collapsed")
     
     # --- Navigation ---
