--- requirements copy.txt
@@ -1,8 +1,0 @@
-streamlit
-pypdf
-mammoth
-python-pptx
-python-docx
-python-dotenv
-tiktoken
-openai>=1.0.0
